Refactored views.py into views/ dir. Separated views into multiple files. Fixed imports.
